custom-elements/custom-element-lifecycle/enqueuing-and-invoking-callbacks/invoke-callbacks.html

<!DOCTYPE html>
<html>
<head>
<title>Invoke CALLBACK with ELEMENT as callback this value and, if present, using string values in CALLBACK as arguments</title>
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Invoke CALLBACK with ELEMENT as callback this value and, if present, using string values in CALLBACK as arguments.">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#enqueuing-and-invoking-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var callbacks = newCustomElementWithCallbacks(doc, 'x-a');

    doc.body.innerHTML = '<x-a id="x-a"></x-a>';
    assert_equals(doc.querySelector('#x-a'), callbacks.createdCallbackThis,
        '\'this\' value of the created callback should be the custom element');
}, 'Test \'this\' value inside created callback.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbacks = newCustomElementWithCallbacks(doc, 'x-element');
    assert_equals(doc.querySelector('#x-element'), callbacks.attachedCallbackThis,
        '\'this\' value of the attached callback should be the custom element');
}, 'Test \'this\' value inside attached callback.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbacks = newCustomElementWithCallbacks(doc, 'x-element');

    var customElement = doc.querySelector('#x-element');
    doc.body.removeChild(customElement);
    assert_equals(customElement, callbacks.detachedCallbackThis,
        '\'this\' value of the detached callback should be the custom element');
}, 'Test \'this\' value inside detached callback.');


test(function() {
    var doc = newHTMLDocument();
    var callbacks = newCustomElementWithCallbacks(doc, 'x-b');
    var customElement = new callbacks.GeneratedConstructor();

    customElement.setAttribute('class', 'someClass');
    assert_equals(customElement, callbacks.attributeChangedCallbackThis,
        '\'this\' value of the attributeChanged callback should be the custom element');
}, 'Test \'this\' value inside attributeChanged callback.');
</script>
</body>
</html>
